requirements.txt
llama-index
langchain 
streamlit
python-dotenv
